requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
